% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classification_report}
\alias{classification_report}
\title{Provides qualitative insights about prediction performances of the logistic
regression model}
\usage{
classification_report(model, data)
}
\arguments{
\item{data}{A data frame containing the qualitative diagnosis elements of several patients.
It can be either the training or the test set, depending on what kind of performances
the user wants to compute.}
}
\value{
A list of two elements giving insights about the performances. The first
one is te confusion matrix, and the second one is the error, expressed in percentages.
}
\description{
Provides qualitative insights about prediction performances of the logistic
regression model
}
\details{
The error is computed using the following formula:\eqn{\text{err} = \dfrac{TP + TN}{P + N}},
where \eqn{TP} and \eqn{TN} are respectively the numbers of true positives and true
negatives, and \eqn{P} and \eqn{N} are respectively the numbers of positives and
negatives samples (whether they are well predicted or not).
}
\examples{
# Here is a whole pipeline through the predictions performances
# Load data
data(patients)
patients <- preprocess_data(patients)
# Train-test split
split <- train_test_split(patients)
train <- split[[1]]
test <- split[[2]]
# Load model and compute the performances
model <- fit_logreg(train)
classification_report(model, train)
classification_report(model, test)
}
