% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert2binary}
\alias{convert2binary}
\title{Convert "yes" or "no" input to a binary input.}
\usage{
convert2binary(response)
}
\arguments{
\item{response}{The response being either "yes" or "no".}
}
\value{
Either 0 or 1
}
\description{
Convert "yes" or "no" input to a binary input.
}
