% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gender2letter}
\alias{gender2letter}
\title{Convert "male" or "female" input into a letter.}
\usage{
gender2letter(gender)
}
\arguments{
\item{gender}{The response being either "male" or "female", with or without
a capital letter.}
}
\value{
Either 'F' or 'M'
}
\description{
Convert "male" or "female" input into a letter.
}
