% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_new_patient}
\alias{get_new_patient}
\title{Loads a new patient data in a data frame}
\usage{
get_new_patient(
  gender,
  age,
  smoking,
  fingers,
  anxiety,
  peers,
  chronic,
  fatigue,
  allergy,
  wheezing,
  alcohol,
  coughing,
  breath,
  swallow,
  chest
)
}
\arguments{
\item{gender}{either 'F' or 'M' for female or male}

\item{age}{an integer giving the age of the patient}

\item{smoking}{either 1 or 0 depending on whether the patient regularly smokes or not.}

\item{fingers}{either 1 or 0 depending on whether the patient has yellow fingers or not.}

\item{anxiety}{either 1 or 0 depending on whether the patient experiences anxiety or not.}

\item{peers}{either 1 or 0 depending on whether the patient experiences peer pressure or not.}

\item{chronic}{either 1 or 0 depending on whether the patient has chronic disease or not.}

\item{fatigue}{either 1 or 0 depending on whether the patient experiences fatigue or not.}

\item{allergy}{either 1 or 0 depending on whether the patient has allergies or not.}

\item{wheezing}{either 1 or 0 depending on whether the patient regularly wheezes or not.}

\item{alcohol}{either 1 or 0 depending on whether the patient has a regular alcohol consumption or not.}

\item{coughing}{either 1 or 0 depending on whether the patient regularly coughs or not.}

\item{breath}{either 1 or 0 depending on whether the patient has a short breath or not.}

\item{swallow}{either 1 or 0 depending on whether the patient has swallowing difficulties or not.}

\item{chest}{either 1 or 0 depending on whether the patient experience chest pain or not.}
}
\value{
A data frame containing all the patient information.
}
\description{
Loads a new patient data in a data frame
}
\details{
The ouput dataframe is already pre-processed as it is done in the `preprocess_data()` function.
}
