% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_summary}
\alias{get_summary}
\title{Display a summary of the dataset provided by the user.}
\usage{
get_summary(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Display a summary of the dataset provided by the user.
}
