% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{outcome}
\alias{outcome}
\title{Returns a prediction of whether a new patient has lung cancer}
\usage{
outcome(model, new_patient)
}
\arguments{
\item{model}{The logistic regression model}

\item{new_patient}{The new patient data loaded in a data frame.}
}
\value{
A sentence giving the prediction of the outcome for this specific patient, along with
the probability given by the model.
}
\description{
Returns a prediction of whether a new patient has lung cancer
}
\examples{
# Here is a whole pipeline from the model to the prediction on a new patient:
data(patients)
patients <- preprocess_data(patients)
model <- fit_logreg(patients)
new_patient <- get_new_patient('F', 24, 0, 1, 1, 1, 0, 0, 0, 1, 0, 1, 1, 0, 1)
outcome(model, new_patient)
}
