% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Pre-process the `patients` dataset}
\usage{
preprocess_data(patients)
}
\arguments{
\item{patients}{Expected to be the `patients` dataset}
}
\value{
The pre-processed `patients` dataset
}
\description{
Pre-process the `patients` dataset
}
\details{
This function performs several pre-processing steps on the `patients` dataset:
- considers categorical columns are considered the right way (useful for Max users)
- change the original encoding of the variables to assign 0 to `NO` and 1 to `YES`.
- binarizes the outcome LUNG_CANCER and makes it categorical in order
to prepare the fit of the logistic regression model.
}
