% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Pre-process the `patients` dataset to assign the right type to categorical columns}
\usage{
preprocess_data(patients)
}
\arguments{
\item{data}{The `patients` dataset}
}
\value{
The pre-processed `patients` dataset
}
\description{
Pre-process the `patients` dataset to assign the right type to categorical columns
}
\details{
This function allows the user to pre-process the patient dataset so that the categorical
columns are considered the right way. This is particularly useful for Mac users, because
it may happen that RStudio on Mac does not recognized such qualitative columns as categorical.
}
