% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{train_test_split}
\alias{train_test_split}
\title{Split the data into a train and a test set}
\usage{
train_test_split(patients, test_size = 0.2)
}
\arguments{
\item{test_size}{A float between 0 and 1 that indicates the proportion of data to sample in the test set. Default value = 0.2.}

\item{data}{A data frame containing the qualitative diagnosis elements of several patients.}
}
\value{
A list of 2 elements: data_train and data_test.
}
\description{
Split the data into a train and a test set
}
\details{
This function returns a random split of the dataset into two datasets, train
and test, regarding the desired test size.
}
\examples{
# Using default proportions
split <- train_test_split(patients)
train = split[[1]]
test = split[[2]]

# Using custom proportions
split <- train_test_split(patients, 0.3)
train = split[[1]]
test = split[[2]]
}
